buggy gameselector in its own file
