+++ client/src/App.js
@@ -3,15 +3,14 @@ import './App.css';
 
 import {Layer, Rect, Text, Stage} from 'react-konva';
 import PropTypes from 'prop-types';
+import { HashRouter, Route } from 'react-router-dom'
 
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import AppBar from 'material-ui/AppBar';
 import Toolbar from 'material-ui/Toolbar';
 import Typography from 'material-ui/Typography';
-import TextField from 'material-ui/TextField';
-import Button from 'material-ui/Button';
-import Paper from 'material-ui/Paper';
-import Grid from 'material-ui/Grid';
+
+import GameSelector from './GameSelector'
 
 export const CONSTANTS = {
   WEBSOCKET: "ws://"+window.location.host+"/websocket/gameview",
@@ -30,10 +29,6 @@ const SUITS = {
   PURPLE: '#F0C',
   BLACK: '#333'
 };
-
-const style = {
-  margin: 12
-}
 
 class Card extends Component {
 
@@ -220,41 +215,7 @@ class Game extends Component {
   };
 };
 
-/**
- * First point of contact, user -> new game or join game
- */
-const paper_style = {
-  padding: 6
-}
-class GameSelector extends Component {
-  render() {
-    return (
-      <Grid container spacing="24">
-        <Grid item xs={1}></Grid>
-        <Grid item xs={5}>
-          <Paper zDepth="2" style={paper_style}>
-            <h3>New Game</h3>
-            <TextField
-              defaultValue="3"
-              label="Number of players"
-            />
-            <Button raised color="primary" style={style}>Start</Button>
-          </Paper>
-        </Grid>
-        <Grid item xs={5}>
-          <Paper zDepth="2" style={paper_style}>
-            <h3>Join Game</h3>
-            <TextField
-              label="Game ID"
-            />
-            <Button raised color="primary" style={style}>Join</Button>
-          </Paper>
-        </Grid>
-        <Grid item xs={1}></Grid>
-      </Grid>
-    )
-  }
-}
+
 
 class App extends Component {
   render() {
@@ -268,7 +229,12 @@ class App extends Component {
               </Typography>
             </Toolbar>
           </AppBar>
-          <div id="content"><GameSelector /></div>
+          <HashRouter>
+            <div id="content">
+              <Route path="/start" component={GameSelector} />
+              <Route path="/gameview" component={Game} />
+            </div>
+          </HashRouter>
           {/* <Game /> */}
         </div>
       </MuiThemeProvider>
